apps/api/workspace/auth/flows/flowExecutor.ts

// flowExecutor.ts
import { FLOWS } from './flowsConfig';
import { useFlowStore } from './flowStore';

export const startFlow = (flowId: string) => {
  const { setFlow, setStep, setParams } = useFlowStore.getState();
  const flow = FLOWS[flowId];
  if (!flow) throw new Error(`Flow with ID \${flowId} not found`);

  setFlow(flowId);
  setStep(flow.steps[0].name);
  executeStep(flow.steps[0].name);
};

export const executeStep = (stepName: string) => {
  const { currentFlow, setStep } = useFlowStore.getState();

  if (!currentFlow) throw new Error('No flow is currently running');

  const flow = FLOWS[currentFlow];
  const step = flow.steps.find((s) => s.name === stepName);
  if (!step) throw new Error(`Step \${stepName} not found in flow \${currentFlow}`);

  // Execute the action for the current step
  const action = actions[step.action];
  if (action) {
    action();
  }

  // Move to the next step if there is one
  const nextStepIndex = flow.steps.findIndex((s) => s.name === stepName) + 1;
  if (nextStepIndex < flow.steps.length) {
    const nextStep = flow.steps[nextStepIndex];
    setStep(nextStep.name);
    executeStep(nextStep.name);
  }
};

const actions = {
  promptLogin: () => {
    console.log('Prompting user to log in...');
    // Implement the actual login prompt logic here
  },
  performLogin: () => {
    console.log('Performing login...');
    // Implement the actual login logic here
  },
  redirectToOrigin: () => {
    console.log('Redirecting to origin...');
    const { params } = useFlowStore.getState();
    // Implement redirection logic here, using params.originUrl
  },
  promptMembership: () => {
    console.log('Prompting user to upgrade membership...');
    // Implement the actual membership prompt logic here
  },
  performMembershipUpgrade: () => {
    console.log('Performing membership upgrade...');
    // Implement the actual membership upgrade logic here
  },
};
